Abort "threads" if slow computations are detected
Because code is currently interpreted in the UI thread, the whole UI
freezes if there are slow computations, e.g. the user accidentally
writes an infinite loop without a call to 'sleep'. To prevent this, I
Introduced a simple detection of slow computations.

The "threads" are aborted by raising an exception if the duration of the
computation is nearing the "look-ahead time" of the sequencer. Such
computations would in any case cause the music to start "stuttering". If
expensive calculation is required, one way to allow it is to increase
the lookahead time of the sequencer.